// src/views/DesignEditor/utils/enum.ts
export enum SLIDER_TYPE {
  SIZE = "size",
  LINE_HEIGHT = "lineHeight",
  LETTER_SPACING = "letterSpacing",
  BRIGHTNESS = "BRIGHTNESS",
  CONTRAST = "CONTRAST",
  SATURATION = "SATURATION",
  HUE = "HUE",
  OPACITY = "OPACITY",
  BLUR = "BLUR",
  HIGHLIGHT = "HIGHLIGHT",
  LOWLIGHT = "LOWLIGHT",
  TEMPERATURE = "TEMPERATURE",
  BANDW = "BANDW",
  NOIR = "NOIR",
  FADE = "FADE",
  MONO = "MONO",
  A2I = "A2I",
  CITY = "CITY",
  BLISS = "BLISS",
  TONAL = "TONAL",
  VINTAGE = "VINTAGE",
  HDR = "HDR",
  LOMO = "LOMO",
  MATTE = "MATTE",
  FILM = "FILM",
  VIBRANT = "VIBRANT",
  COOLTONE = "COOLTONE",
}